+++ caf-propertysource-spring/pom.xml
@@ -26,7 +26,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>caf-propertyresolver-spring</artifactId>
+    <artifactId>caf-propertysource-spring</artifactId>
     
     <dependencies>
         <dependency>
